NFC: Move StrEnum classes to enums.py
