+++ src/apps/audit/__init__.py
@@ -1,4 +1,5 @@
-from apps.audit.domain import AuditEvent, EventAction, EventOutcome
+from apps.audit.domain import AuditEvent
+from apps.audit.enums import EventAction, EventOutcome
 from apps.audit.service import log
 
 __all__ = ["AuditEvent", "EventAction", "EventOutcome", "log"]
